refactor path integration code - add modules
